Make asserts and promises play together (#36)
* Oooops.  Didn't commit some of the changes from assert to assert.equal.  How'd that happen!?

* Use mocha to handle promises in unit tests.

Fixes #30

Before this change, if an assert in a promise fails, mocha does not print out the failure, instead, the test will timeout which is the only indication that something is wrong.

Also, if the test doesn't have a done() function, and an assert fails in a promise, ther is no indication that anything is wrong.

This change switches from using the done() callback for async tests to returning promises to mocha.  The big advantage of this change is that if an assert fails in a promise, mocha will no display info about the failed assert as one would expect.  Another significant advantage is that it is no longer the test writers responsibility to handle rejected promises.  Mocha will complain about rejected promises just like it should.

The test writer does need to be careful that they return all promises, but test writers should be testing that a failed test behaves as they expect anyway, so this shouldn't be a practicle problem.

The changes in the commit are to remove the done parameter from each test, and to make sure that all async functions ultimately return a promuise to the test runner.

In addition, there are two commented out tests, those have been update to use const, assert.equal and in one case the mode of commenting is changed from /* */ to // just to make iut easier for me to comment and uncomment which will ensure that they stay up to date :).