Add support for Aftersave (#65)
We have an afterSave hook for a Parse object, which we'd like to test without directly calling it (and to make the tests closer to reality, meaning that the afterSave hook should run after saving). In order to do this, I've added afterSave hook support to the parse-mockDB.

The existing tests for behavesLikeParseObjectOnBeforeSave are re-used, except that the afterSavePromise mutates a local boolean as to confirm that it ran and succeeded.