Return full saved object in afterSave hooks
In 78296a03f1, a bug was introduced where we would invoke afterSave
hooks with only the subset of fields that was being directly returned
from the invocation to `save()`. This doesn't match Parse Server's
behavior, or the tests.

Of note: commit 78296a03f1 actually broke tests that were originally
written in commit b42ed240d1 (it "rejects the save if there is a
problem"), because the "error" field wasn't being passed down into the
after save hook, preventing it from triggering the error state. The test
was updated so that it would pass (it "...is saved even if there was a
problem").

Because this new test does describe how I expect Parse Server to behave,
I've updated MockDB to match this specification instead of spuriously
passing. I also did a little bit of cleanup on the tests themselves, to
simplify and read more nicely.